+++ src/richie/plugins/section/models.py
@@ -6,6 +6,7 @@ from django.utils.text import Truncator
 from django.utils.translation import gettext_lazy as _
 
 from cms.models.pluginmodel import CMSPlugin
+from djangocms_attributes_field.fields import AttributesField
 
 from .defaults import SECTION_TEMPLATES
 
@@ -26,6 +27,11 @@ class Section(CMSPlugin):
         choices=SECTION_TEMPLATES,
         default=SECTION_TEMPLATES[0][0],
         help_text=_("Optional template for custom look."),
+    )
+    attributes = AttributesField(
+        verbose_name=_("Attributes"),
+        blank=True,
+        excluded_keys=["href", "target"],
     )
 
     def __str__(self):