+++ CHANGELOG.md
@@ -23,6 +23,9 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Fix schema.org related "pt_effort" string computation when effort is not set
 - Use `$r-course-subheader-aside` to define subheader aside column width
+### Changed
+
+- Add attributes to the Section plugin
 
 ## [2.8.1] - 2021-09-28